+++ board_files/include/output/posting_form.php
@@ -34,7 +34,7 @@ function nel_render_posting_form($board_id, $render, $response_to, $dotdot = nul
     $new_post_element = $posting_form->doXPathQuery(".//input[@name='new_post[post_info][response_to]']", $posting_form)->item(0);
     $new_post_element->extSetAttribute('value', $response_to);
 
-    if(!nel_sessions()->sessionIsIgnored())
+    if(!nel_sessions()->sessionIsIgnored('render'))
     {
         $post_action = $posting_form->doXPathQuery(".//input[@name='action']", $posting_form)->item(0);
         $post_action->extSetAttribute('value', 'new-staff-post');